SOMEONE COULD HAVE TOLD GOD THAT THE ELF MARCH DIDN'T MATCH UP AT ALL TO THE CUTSCENE
it even cuts off right as the build section of the song ends. So I just cut the whole build section off and it has no intro anymore. It just goes straight into the march. Also I slowed it back down because it didn't work with the footsteps anyways so I wouldn't really call it a march. Seems like a bad term all around.